+++ doc/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 Unreleased
 ----------
 
+- Add a stopping criterion and parameter tuning heuristic for Huber robust mean
+    estimator.
 - Add `CLARA` (Clustering for Large Applications) which extends k-medoids to
     be more scalable using a sampling approach.
     [`#83 <https://github.com/scikit-learn-contrib/scikit-learn-extra/pull/83>`_].